refactor: reorganise helper functions
