+++ packages/create-solid/src/index.ts
@@ -5,8 +5,9 @@ import packageJson from "../package.json" with { type: "json" };
 import { defineCommand, runMain } from "citty";
 import { createVanilla, createVanillaJS } from "./create-vanilla";
 import * as p from "@clack/prompts";
-import { cancelable, getTemplatesList, StartTemplate, VanillaTemplate } from "./helpers";
+import { cancelable, getTemplatesList, spinnerify, StartTemplate, VanillaTemplate } from "./helpers";
 import { createStart } from "./create-start";
+import { create } from "./create";
 intro(`\n${color.bgCyan(color.black(` Create-Solid v${packageJson.version}`))}`);
 
 const main = defineCommand({
@@ -58,9 +59,17 @@ const main = defineCommand({
 		isJS = !(await cancelable(p.confirm({ message: "Use Typescript?" })))
 
 		if (isStart) {
-			createVanilla({ template: template as StartTemplate, destination: projectName }, isJS);
+			await spinnerify({
+				startText: "Creating project",
+				finishText: "Project created 🎉",
+				fn: () => createStart({ template: template as StartTemplate, destination: projectName }, isJS),
+			})
 		} else {
-			createVanilla({ template: template as VanillaTemplate, destination: projectName }, isJS);
+			await spinnerify({
+				startText: "Creating project",
+				finishText: "Project created 🎉",
+				fn: () => createVanilla({ template: template as VanillaTemplate, destination: projectName }, isJS),
+			})
 		}
 	}
 })
