+++ packages/create-solid/src/utils/file-system.ts
@@ -1,0 +1,18 @@
+import { copyFileSync, Dirent, mkdirSync, readdirSync, writeFileSync } from "node:fs";
+import { readFile, rm } from "node:fs/promises";
+import { join, resolve } from "node:path";
+import { transform } from "sucrase";
+
+export const recurseFiles = (startPath: string, cb: (file: Dirent, startPath: string) => void) => {
+    startPath = resolve(startPath);
+
+    const files = readdirSync(startPath, { withFileTypes: true });
+
+    for (const file of files) {
+        cb(file, startPath);
+    }
+};
+
+export const readFileToString = async (path: string) => {
+    return (await readFile(path)).toString();
+};
